Fix warnings
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9230 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/a52.c
@@ -128,7 +128,7 @@ extern char iend[];
 /* this is the codec entry point */
 enum codec_status codec_start(struct codec_api *api)
 {
-    long n;
+    size_t n;
     unsigned char *filebuf;
     int sample_loc;
     int retval;
